+++ app/lib/importers/all_organisations.rb
@@ -1,7 +1,7 @@
 module Importers
   class AllOrganisations
     def run
-      Collectors::Organisations.new.find_each do |attributes|
+      OrganisationsService.new.find_each do |attributes|
         organisation = ::Organisation.find_or_create_by(slug: attributes.fetch(:slug))
         organisation.assign_attributes(attributes)
         organisation.save!
